+++ src/assets/icons.ts
