+++ src/components/confirmDeleteComponent.tsx
@@ -1,6 +1,6 @@
 import { Fragment, h } from "preact";
 import t from "src/l10n";
-import ConfirmDeleteModal from "../ui/confirmDeleteModal";
+import ConfirmDeleteModal from "../settings/confirmDeleteModal";
 
 export function confirmDeleteComponent({
 	modal,
