+++ src/components/settingTabComponent.tsx
@@ -1,7 +1,7 @@
 import { Fragment, h } from "preact";
 import t from "src/l10n";
-import { ObsidianIcon, isTriggerInConflicts } from "src/util";
-import { buildQueryPattern } from "src/search";
+import { ObsidianIcon, isTriggerInConflicts } from "src/utils/util";
+import { buildQueryPattern } from "src/utils/search";
 import SlashCommanderPlugin from "../main";
 import CommandViewer from "./commandViewerComponent";
 import {
