+++ src/components/suggestionComponent.tsx
@@ -1,7 +1,7 @@
 import { Fragment, h } from "preact";
 import SlashCommanderPlugin from "src/main";
 import { CommandIconPair } from "src/types";
-import { getCommandFromId, getCommandSourceName, isCommandNameUnique, ObsidianIcon } from "src/util";
+import { getCommandFromId, getCommandSourceName, isCommandNameUnique, ObsidianIcon } from "src/utils/util";
 
 interface SuggestionProps {
 	plugin: SlashCommanderPlugin;
