+++ src/settings/addCommandModal.ts
