+++ src/settings/confirmDeleteModal.ts
